Delete DominoChain class, move all functions and LinkedList<Domino> dominoChain to the Game class. Delete all useless lines.
