+++ src/logic/dominoes/CommandHistory.java
@@ -1,5 +1,4 @@
 package logic.dominoes;
-import logic.dominoes.DominoComand;
 
 import java.util.Stack;
