+++ src/logic/players/ComputerCreator.java
@@ -1,8 +1,5 @@
 package logic.players;
 
-import logic.players.Computer;
-import logic.players.Player;
-import logic.players.PlayerCreator;
 
 public class ComputerCreator extends PlayerCreator {
